# Slotted array faster defragmentation