--- a/src/Paprika.Benchmarks/SlottedArrayBenchmarks.cs
+++ b/src/Paprika.Benchmarks/SlottedArrayBenchmarks.cs
@@ -1,5 +1,3 @@
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 using BenchmarkDotNet.Attributes;
 using Paprika.Crypto;
 using Paprika.Data;
@@ -11,7 +9,12 @@ namespace Paprika.Benchmarks;
 [DisassemblyDiagnoser(maxDepth: 2)]
 public class SlottedArrayBenchmarks
 {
+    // defragmentation
     private readonly byte[] _defragmentation = new byte[Page.PageSize];
+    private readonly byte[] _defragmentationCopy = new byte[Page.PageSize];
+    private readonly byte[] _defragmentationValue = new byte[64];
+    private readonly ushort _writtenTo;
+
     private readonly byte[] _writtenLittleEndian = new byte[Page.PageSize];
     private readonly byte[] _writtenBigEndian = new byte[Page.PageSize];
     private readonly byte[] _writable = new byte[Page.PageSize];
@@ -27,41 +30,62 @@ public class SlottedArrayBenchmarks
 
     public SlottedArrayBenchmarks()
     {
-        var little = new SlottedArray(_writtenLittleEndian);
-        var big = new SlottedArray(_writtenBigEndian);
-
-        Span<byte> key = stackalloc byte[4];
-
         // Big and small endian tests
-        while (true)
         {
-            WriteInt32LittleEndian(key, _to);
-            if (little.TrySet(NibblePath.FromKey(key), key) == false)
-            {
-                // filled
-                break;
-            }
+            var little = new SlottedArray(_writtenLittleEndian);
+            var big = new SlottedArray(_writtenBigEndian);
 
-            WriteInt32BigEndian(key, _to);
-            if (big.TrySet(NibblePath.FromKey(key), key) == false)
-            {
-                // filled
-                break;
-            }
+            Span<byte> key = stackalloc byte[4];
 
-            _to++;
+
+            while (true)
+            {
+                WriteInt32LittleEndian(key, _to);
+                if (little.TrySet(NibblePath.FromKey(key), key) == false)
+                {
+                    // filled
+                    break;
+                }
+
+                WriteInt32BigEndian(key, _to);
+                if (big.TrySet(NibblePath.FromKey(key), key) == false)
+                {
+                    // filled
+                    break;
+                }
+
+                _to++;
+            }
         }
 
         // Hash collisions tests
-        var zeroes = NibblePath.FromKey(Keccak.Zero);
-        var hashCollisions = new SlottedArray(_hashCollisions);
-
-        for (var i = 0; i <= HashCollisionsCount; i++)
         {
-            if (!hashCollisions.TrySet(zeroes.SliceTo(i), HashCollisionValue))
+            var zeroes = NibblePath.FromKey(Keccak.Zero);
+            var hashCollisions = new SlottedArray(_hashCollisions);
+
+            for (var i = 0; i <= HashCollisionsCount; i++)
             {
-                throw new Exception($"No place to set hash collision at {i}");
+                if (!hashCollisions.TrySet(zeroes.SliceTo(i), HashCollisionValue))
+                {
+                    throw new Exception($"No place to set hash collision at {i}");
+                }
             }
+        }
+
+        // Defragmentation
+        {
+            var map = new SlottedArray(_defragmentation);
+            ushort i = 0;
+            Span<byte> key = stackalloc byte[2];
+
+            // Set as many as possible
+            while (map.TrySet(NibblePath.FromKey(key), _defragmentationValue))
+            {
+                i++;
+                WriteUInt16LittleEndian(key, i);
+            }
+
+            _writtenTo = i;
         }
     }
 
@@ -240,22 +264,16 @@ public class SlottedArrayBenchmarks
     }
 
     private const int DefragmentOpsCount = 4;
+
     [Benchmark(OperationsPerInvoke = DefragmentOpsCount)]
     public void Defragment_first_key_deleted()
     {
-        var map = new SlottedArray(_defragmentation);
-        map.Clear();
+        _defragmentation.CopyTo(_defragmentationCopy.AsSpan());
 
-        var value = _writtenLittleEndian.AsSpan(64);
-        ushort i = 0;
+        var map = new SlottedArray(_defragmentationCopy);
+
         Span<byte> key = stackalloc byte[2];
-
-        // Set as many as possible
-        while (map.TrySet(NibblePath.FromKey(key), value))
-        {
-            i++;
-            WriteUInt16LittleEndian(key, i);
-        }
+        var i = _writtenTo;
 
         // Delete & defragment
         for (ushort j = 0; j < DefragmentOpsCount; j++)
@@ -266,39 +284,30 @@ public class SlottedArrayBenchmarks
 
             // Encode new key and set
             WriteUInt16LittleEndian(key, i++);
-            map.TrySet(NibblePath.FromKey(key), value);
+            map.TrySet(NibblePath.FromKey(key), _defragmentationValue);
         }
     }
 
     [Benchmark(OperationsPerInvoke = DefragmentOpsCount)]
     public void Defragment_last_key_deleted()
     {
-        var map = new SlottedArray(_defragmentation);
-        map.Clear();
+        _defragmentation.CopyTo(_defragmentationCopy.AsSpan());
 
-        var value = _writtenLittleEndian.AsSpan(64);
-        ushort i = 0;
+        var map = new SlottedArray(_defragmentationCopy);
+
         Span<byte> key = stackalloc byte[2];
-
-        // Set as many as possible
-        while (map.TrySet(NibblePath.FromKey(key), value))
-        {
-            i++;
-            WriteUInt16LittleEndian(key, i);
-        }
-
-        i--; // to go back to the existing key
+        var last = (ushort)(_writtenTo - 1);
 
         // Delete & defragment
-        for (var j = 0; j < DefragmentOpsCount; j++)
+        for (ushort j = 0; j < DefragmentOpsCount; j++)
         {
-            // Delete first existing, the last one
-            WriteUInt16LittleEndian(key, i);
+            // Delete first
+            WriteUInt16LittleEndian(key, last);
             map.Delete(NibblePath.FromKey(key));
 
             // Encode new key and set
-            WriteUInt16LittleEndian(key, i++);
-            map.TrySet(NibblePath.FromKey(key), value);
+            WriteUInt16LittleEndian(key, last++);
+            map.TrySet(NibblePath.FromKey(key), _defragmentationValue);
         }
     }
 }