This PR changes the `SlottedArray.Defragment` method that is clearly visible in traces and previously was copying data twice, once from the array to a buffer, then buffer to an array. This PR removes the need of the auxiliary buffer which required renting & releasing but also introduced one more copy.

The benchmarks below does not show the gains that are observed. P95 of the block application to the database dropped by ~20% which is a lot.

### Benchmark 

#### Before

| Method                       | Mean      | Error    | StdDev   | Code Size |
|----------------------------- |----------:|---------:|---------:|----------:|
| Defragment_first_key_deleted | 372.76 ns | 3.585 ns | 3.178 ns |   7,438 B |
| Defragment_last_key_deleted  |  30.19 ns | 0.473 ns | 0.443 ns |   5,609 B |

#### After

| Method                       | Mean      | Error    | StdDev   | Code Size |
|----------------------------- |----------:|---------:|---------:|----------:|
| Defragment_first_key_deleted | 258.77 ns | 2.896 ns | 2.567 ns |   5,553 B |
| Defragment_last_key_deleted  |  29.75 ns | 0.274 ns | 0.256 ns |   5,607 B |